Login con validaciones y cerrar sesion

+++ frontend/src/app/estructura/nav/nav.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-
 import { Router, RouterModule } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-nav',
@@ -23,12 +23,24 @@ export class Nav {
   }
 
   cerrar() {
-    sessionStorage.setItem("id", "");
-    sessionStorage.setItem("email", "");
-    sessionStorage.setItem("nombres", "");
-    sessionStorage.removeItem("apellidos");
-    sessionStorage.setItem("rol", "");
-    this.router.navigate(['login']);
-  }
+    Swal.fire({
+      title: "¿Desea cerrar sesión?",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, cerrar sesión",
+      cancelButtonText: "Cancelar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        sessionStorage.setItem("id", "");
+        sessionStorage.setItem("email", "");
+        sessionStorage.setItem("nombres", "");
+        sessionStorage.removeItem("apellidos");
+        sessionStorage.setItem("rol", "");
 
+        this.router.navigate(['login']);
+      }
+    });
+  }
 }
